NewFuctions

+++ app/Http/Controllers/Admin/ProductsController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\Product;
+use App\Models\User;
+use App\Scopes\ActiveStatusScope;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
-
 
 class ProductsController extends Controller
 {
@@ -19,15 +21,20 @@ class ProductsController extends Controller
      */
     public function index()
     {
+        $this->authorize('view-any', Product::class);
 
-        
-        $products = Product::join('categories' , 'categories.id' ,'=' ,'products.category_id')
-        ->select([
-            'products.*',
-            'categories.name as category_name'
-        ])->paginate(10);
-        return view('admin.products.index')
-        ->withProducts($products);
+        $products = Product::withoutGlobalScopes([ActiveStatusScope::class])
+            ->join('categories', 'categories.id', '=', 'products.category_id')
+            ->select([
+                'products.*',
+                'categories.name as category_name',
+            ])
+            ->paginate(15);
+
+
+        return view('admin.products.index', [
+            'products' => $products,
+        ]);
     }
 
     /**
@@ -37,12 +44,17 @@ class ProductsController extends Controller
      */
     public function create()
     {
-          $categories = Category::all();
-          return view('admin.products.create', [
-          'categories' => $categories,
-          'product' => new Product(),
-          ]);
+        /*if (!Gate::allows('products.create')) {
+            abort(403);
+        }*/
+        $this->authorize('create', Product::class);
 
+        $categories = Category::pluck('name', 'id');
+
+        return view('admin.products.create', [
+            'categories' => $categories,
+            'product' => new Product(),
+        ]);
     }
 
     /**
@@ -53,16 +65,20 @@ class ProductsController extends Controller
      */
     public function store(Request $request)
     {
+        /*if (!Gate::allows('products.create')) {
+            abort(403);
+        }*/
+        $this->authorize('create', Product::class);
+
         $request->validate(Product::validateRules());
 
-        $request->merge([
-            'slug' => Str::slug($request->post('name'))
-        ]);
-
+        /*$request->merge([
+            'slug' => Str::slug($request->post('name')),
+        ]);*/
         $product = Product::create($request->all());
 
-          return redirect()->route('products.index')
-          ->with('success', "Product ($product->name) created.");
+        return redirect()->route('products.index')
+            ->with('success', "Product ($product->name) created.");
     }
 
     /**
@@ -73,9 +89,13 @@ class ProductsController extends Controller
      */
     public function show($id)
     {
-      $product = Product::findOrFail($id);
-      return view('admin.products.show')
-      ->withProducts($products);
+        $product = Product::withoutGlobalScope('active')->findOrFail($id);
+
+        $this->authorize('view', $product);
+
+        return view('admin.products.show', [
+            'product' => $product,
+        ]);
     }
 
     /**
@@ -86,10 +106,12 @@ class ProductsController extends Controller
      */
     public function edit($id)
     {
-       $product = Product::findOrFail($id);
+        $product = Product::withoutGlobalScope('active')->findOrFail($id);
+        $this->authorize('update', $product);
+
         return view('admin.products.edit', [
             'product' => $product,
-            'categories' => Category::all(),
+            'categories' => Category::withTrashed()->pluck('name', 'id'),
         ]);
     }
 
@@ -100,14 +122,21 @@ class ProductsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Product $product)
     {
-          $product = Product::findOrFail($id);
+        //$product = Product::withoutGlobalScope('active')->findOrFail($id);
+        $this->authorize('update', $product);
 
-        $request->validate( Product::validateRules() );
+        $request->validate(Product::validateRules());
 
         if ($request->hasFile('image')) {
             $file = $request->file('image'); // UplodedFile Object
+            // $file->getClientOriginalName(); // Return file name
+            // $file->getClientOriginalExtension();
+            // $file->getClientMimeType(); // audio/mp3
+            // $file->getType();
+            // $file->getSize();
+
             // Filesystem - Disks
             // local: storage/app
             // public: storage/app/public
@@ -119,13 +148,13 @@ class ProductsController extends Controller
             $request->merge([
                 'image_path' => $image_path,
             ]);
+        }
 
-            $product->update( $request->all() );
+        $product->update($request->all());
 
-            return redirect()->route('products.index')
+        return redirect()->route('products.index')
             ->with('success', "Product ($product->name) updated.");
-       }
-   }
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -135,14 +164,57 @@ class ProductsController extends Controller
      */
     public function destroy($id)
     {
-       $product = Product::findOrFail($id);
+        //Gate::authorize('products.delete');
+
+        //$user = User::find(1);
+        //Gate::forUser($user)->allows('products.delete');
+
+        $product = Product::withoutGlobalScope('active')->findOrFail($id);
+        $this->authorize('delete', $product);
         $product->delete();
 
-        Storage::disk('uploads')->delete($product->image_path);
+        //Storage::disk('uploads')->delete($product->image_path);
         //unlink(public_path('uploads/' . $product->image_path));
 
         return redirect()->route('products.index')
-        ->with('success', "Product ($product->name) deleted.");
+            ->with('success', "Product ($product->name) deleted.");
+    }
 
+    public function trash()
+    {
+        $products = Product::withoutGlobalScope('active')->onlyTrashed()->paginate();
+        return view('admin.products.trash', [
+            'products' => $products,
+        ]);
+    }
+
+    public function restore(Request $request, ?Product $product = null)
+    {
+        if ($product) {
+            //$product = Product::withoutGlobalScope('active')->onlyTrashed()->findOrFail($id);
+            $product->restore();
+
+            return redirect()->route('products.index')
+                ->with('success', "Product ($product->name) restored.");
+        }
+
+        Product::withoutGlobalScope('active')->onlyTrashed()->restore();
+        return redirect()->route('products.index')
+            ->with('success', "All trashed products restored.");
+    }
+
+    public function forceDelete($id = null)
+    {
+        if ($id) {
+            $product = Product::withoutGlobalScope('active')->onlyTrashed()->findOrFail($id);
+            $product->forceDelete();
+
+            return redirect()->route('products.index')
+                ->with('success', "Product ($product->name) deleted forever.");
+        }
+
+        Product::withoutGlobalScope('active')->onlyTrashed()->forceDelete();
+        return redirect()->route('products.index')
+            ->with('success', "All trashed products deleted forever.");
     }
 }
